refactoring lexer
