+++ compiler/src/lexer/lexer.rs
@@ -1,7 +1,6 @@
 use super::token::LexicalErrorKind;
 use crate::code::JarvilCode;
 use crate::error::diagnostics::{Diagnostics, InvalidCharError, NoClosingSymbolError};
-use crate::lexer::helper;
 use crate::lexer::token::CoreToken;
 use crate::lexer::token::Token;
 use std::convert::TryFrom;
@@ -10,12 +9,16 @@ use std::vec;
 use text_size::TextRange;
 use text_size::TextSize;
 
+pub fn is_letter(c: &char) -> bool {
+    c.is_ascii_alphabetic() || (*c == '_')
+}
+
 pub trait Lexer {
     fn tokenize(self, code: &mut JarvilCode) -> (Vec<Token>, Vec<Diagnostics>);
 }
 
 pub struct CoreLexer {
-    pub begin_lexeme: usize,
+    pub index: usize,
     pub line_number: usize,
     pub code_lines: Vec<usize>,
     pub line_start_index: usize,
@@ -35,13 +38,13 @@ impl Lexer for CoreLexer {
             trivia: None,
         });
         let mut eof_trivia_vec: Vec<Token> = vec![];
-        while self.begin_lexeme < code.len() {
+        while self.index < code.len() {
             let token = self.extract_lexeme(code);
             if token.is_trivia() {
                 let mut trivia_vec = vec![token];
                 let mut is_eof: Option<Token> = None;
                 // spin loop to collect other trivia tokens as well
-                while self.begin_lexeme < code.len() {
+                while self.index < code.len() {
                     let token = self.extract_lexeme(code);
                     if token.is_trivia() {
                         trivia_vec.push(token);
@@ -90,7 +93,7 @@ impl Lexer for CoreLexer {
 impl CoreLexer {
     pub fn new() -> Self {
         CoreLexer {
-            begin_lexeme: 0,
+            index: 0,
             line_number: 1,
             code_lines: vec![],
             line_start_index: 0,
@@ -99,113 +102,84 @@ impl CoreLexer {
     }
 
     fn extract_lexeme(&mut self, code: &JarvilCode) -> Token {
-        let begin_lexeme = &mut self.begin_lexeme;
-        let line_number = &mut self.line_number;
-        let code_lines = &mut self.code_lines;
-        let line_start_index = &mut self.line_start_index;
-
-        let start_index = *begin_lexeme;
-        let start_line_number = *line_number;
-        let critical_char = code.get_char(*begin_lexeme);
+        let start_index = self.index;
+        let start_line_number = self.line_number;
+        let critical_char = code.get_char(self.index);
         let core_token = match critical_char {
             '(' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::LPAREN
             }
             ')' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::RPAREN
             }
             '{' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::LBRACE
             }
             '}' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::RBRACE
             }
             '[' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::LSQUARE
             }
             ']' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::RSQUARE
             }
             ';' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::SEMICOLON
             }
             ',' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::COMMA
             }
             '.' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::DOT
             }
-            /*
-            '\t'        =>      {
-                *begin_lexeme = *begin_lexeme + 1;
-                (CoreToken::TAB, String::from("\t"))
-            },
-             */
             '+' => {
-                *begin_lexeme += 1;
+                self.index += 1;
                 CoreToken::PLUS
             }
             '\n' => {
-                code_lines.push(*line_start_index);
-                *line_start_index = *begin_lexeme + 1;
-                *begin_lexeme += 1;
-                *line_number += 1;
+                self.code_lines.push(self.line_start_index);
+                self.line_start_index = self.index + 1;
+                self.index += 1;
+                self.line_number += 1;
                 CoreToken::NEWLINE
             }
-            '/' => helper::extract_slash_prefix_lexeme(
-                begin_lexeme,
-                line_number,
-                code,
-                code_lines,
-                line_start_index,
-            ),
-            '"' => helper::extract_double_quote_prefix_lexeme(
-                begin_lexeme,
-                line_number,
-                code,
-                code_lines,
-                line_start_index,
-            ),
-            '\'' => helper::extract_single_quote_prefix_lexeme(
-                begin_lexeme,
-                line_number,
-                code,
-                code_lines,
-                line_start_index,
-            ),
-            '!' => helper::extract_exclaimation_prefix_lexeme(begin_lexeme, code),
-            ' ' => helper::extract_blank_prefix_lexeme(begin_lexeme, code),
-            '-' => helper::extract_dash_prefix_lexeme(begin_lexeme, code),
-            '*' => helper::extract_star_prefix_lexeme(begin_lexeme, code),
-            '#' => helper::extract_hash_prefix_lexeme(begin_lexeme, code),
-            '=' => helper::extract_equal_prefix_lexeme(begin_lexeme, code),
-            '>' => helper::extract_rbracket_prefix_lexeme(begin_lexeme, code),
-            '<' => helper::extract_lbracket_prefix_lexeme(begin_lexeme, code),
-            ':' => helper::extract_colon_prefix_lexeme(begin_lexeme, code),
+            '/' => self.extract_slash_prefix_lexeme(code),
+            '"' => self.extract_double_quote_prefix_lexeme(code),
+            '\'' => self.extract_single_quote_prefix_lexeme(code),
+            '!' => self.extract_exclaimation_prefix_lexeme(code),
+            ' ' => self.extract_blank_prefix_lexeme(code),
+            '-' => self.extract_dash_prefix_lexeme(code),
+            '*' => self.extract_star_prefix_lexeme(code),
+            '#' => self.extract_hash_prefix_lexeme(code),
+            '=' => self.extract_equal_prefix_lexeme(code),
+            '>' => self.extract_rbracket_prefix_lexeme(code),
+            '<' => self.extract_lbracket_prefix_lexeme(code),
+            ':' => self.extract_colon_prefix_lexeme(code),
             c => {
                 let token: CoreToken;
-                if helper::is_letter(&c) {
-                    token = helper::extract_letter_prefix_lexeme(begin_lexeme, code);
+                if is_letter(&c) {
+                    token = self.extract_letter_prefix_lexeme(code);
                 } else if c.is_ascii_digit() {
-                    token = helper::extract_digit_prefix_lexeme(begin_lexeme, code);
+                    token = self.extract_digit_prefix_lexeme(code);
                 } else {
-                    *begin_lexeme += 1;
+                    self.index += 1;
                     token = CoreToken::LEXICAL_ERROR(LexicalErrorKind::InvalidChar);
                 }
                 token
             }
         };
-        let end_index = *begin_lexeme;
-        let end_line_number = *line_number;
+        let end_index = self.index;
+        let end_line_number = self.line_number;
         let token = Token {
             line_number: start_line_number,
             core_token: core_token.clone(),
@@ -231,6 +205,380 @@ impl CoreLexer {
             _ => {}
         }
         token
+    }
+
+    // ' ' -> '...'
+    pub fn extract_blank_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            if next_char != ' ' {
+                self.index = forward_lexeme;
+                return CoreToken::BLANK;
+            }
+            forward_lexeme += 1;
+        }
+        self.index = forward_lexeme;
+        CoreToken::BLANK
+    }
+
+    // - -> -, ->
+    pub fn extract_dash_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '>' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::RIGHT_ARROW
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::DASH
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::DASH
+        }
+    }
+
+    // * -> *, **
+    pub fn extract_star_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '*' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::DOUBLE_STAR
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::STAR
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::STAR
+        }
+    }
+
+    // / -> /, /*, //
+    pub fn extract_slash_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        let mut state: usize = 0;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match state {
+                0 => match next_char {
+                    '/' => {
+                        state = 1;
+                    }
+                    '*' => {
+                        state = 2;
+                    }
+                    _ => {
+                        self.index += 1;
+                        return CoreToken::SLASH;
+                    }
+                },
+                1 => match next_char {
+                    '\n' => {
+                        self.index = forward_lexeme;
+                        return CoreToken::SINGLE_LINE_COMMENT;
+                    }
+                    _ => {}
+                },
+                2 => match next_char {
+                    '*' => {
+                        state = 3;
+                    }
+                    '\n' => {
+                        self.code_lines.push(self.line_start_index);
+                        self.line_number += 1;
+                        self.line_start_index = forward_lexeme + 1;
+                    }
+                    _ => {}
+                },
+                3 => match next_char {
+                    '/' => {
+                        self.index = forward_lexeme + 1;
+                        return CoreToken::BLOCK_COMMENT;
+                    }
+                    _ => {
+                        state = 2;
+                    }
+                },
+                _ => {
+                    unreachable!("any state other than 0, 1, 2 and 3 is not reachable")
+                }
+            }
+            forward_lexeme += 1;
+        }
+        match state {
+            0 => {
+                self.index += 1;
+                CoreToken::SLASH
+            }
+            1 => {
+                self.index = forward_lexeme;
+                CoreToken::SINGLE_LINE_COMMENT
+            }
+            2 => {
+                self.index = forward_lexeme;
+                CoreToken::LEXICAL_ERROR(LexicalErrorKind::NoClosingSymbols("*/"))
+            }
+            3 => {
+                self.index = forward_lexeme;
+                CoreToken::LEXICAL_ERROR(LexicalErrorKind::NoClosingSymbols("*/"))
+            }
+            _ => unreachable!("any state other than 0, 1, 2 and 3 is not reachable"),
+        }
+    }
+
+    // # -> #...\n
+    pub fn extract_hash_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '\n' => {
+                    self.index = forward_lexeme;
+                    return CoreToken::SINGLE_LINE_COMMENT;
+                }
+                _ => {}
+            }
+            forward_lexeme += 1;
+        }
+        self.index = forward_lexeme;
+        CoreToken::SINGLE_LINE_COMMENT
+    }
+
+    // = -> =, ==
+    pub fn extract_equal_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '=' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::DOUBLE_EQUAL
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::EQUAL
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::EQUAL
+        }
+    }
+
+    // > -> >, >=
+    pub fn extract_rbracket_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '=' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::GREATER_EQUAL
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::RBRACKET
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::RBRACKET
+        }
+    }
+
+    // < -> <, <=
+    pub fn extract_lbracket_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '=' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::LESS_EQUAL
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::LBRACKET
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::LBRACKET
+        }
+    }
+
+    // ! -> !=
+    pub fn extract_exclaimation_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '=' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::NOT_EQUAL
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::LEXICAL_ERROR(LexicalErrorKind::InvalidChar)
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::LEXICAL_ERROR(LexicalErrorKind::InvalidChar)
+        }
+    }
+
+    // ' -> '...'
+    pub fn extract_single_quote_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '\'' => {
+                    self.index = forward_lexeme + 1;
+                    return CoreToken::LITERAL;
+                }
+                '\n' => {
+                    self.code_lines.push(self.line_start_index);
+                    self.line_number += 1;
+                    self.line_start_index = forward_lexeme + 1;
+                }
+                _ => {}
+            }
+            forward_lexeme += 1;
+        }
+        self.index = forward_lexeme;
+        CoreToken::LEXICAL_ERROR(LexicalErrorKind::NoClosingSymbols("'"))
+    }
+
+    // " -> "..."
+    pub fn extract_double_quote_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                '"' => {
+                    self.index = forward_lexeme + 1;
+                    return CoreToken::LITERAL;
+                }
+                '\n' => {
+                    self.code_lines.push(self.line_start_index);
+                    self.line_number += 1;
+                    self.line_start_index = forward_lexeme + 1;
+                }
+                _ => {}
+            }
+            forward_lexeme += 1;
+        }
+        self.index = forward_lexeme;
+        CoreToken::LEXICAL_ERROR(LexicalErrorKind::NoClosingSymbols(r#"""#))
+    }
+
+    // letter -> letter((letter|digit|_)*) or keyword or type
+    pub fn extract_letter_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            if is_letter(&next_char) || next_char.is_ascii_digit() {
+                // do nothing
+            } else {
+                let value_iter = code.token_value_as_iter(self.index, Some(forward_lexeme));
+                self.index = forward_lexeme;
+                return CoreToken::token_for_identifier(value_iter);
+            }
+            forward_lexeme += 1;
+        }
+        let value_iter = code.token_value_as_iter(self.index, Some(forward_lexeme));
+        self.index = forward_lexeme;
+        CoreToken::token_for_identifier(value_iter)
+    }
+
+    // digit -> digit((digit)*(.digit(digit*)|empty))
+    pub fn extract_digit_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let mut forward_lexeme = self.index + 1;
+        let mut state: usize = 0;
+        while forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match state {
+                0 => {
+                    if next_char.is_ascii_digit() {
+                        // do nothing
+                    } else if next_char == '.' {
+                        state = 1;
+                    } else {
+                        self.index = forward_lexeme;
+                        return CoreToken::INTEGER;
+                    }
+                }
+                1 => {
+                    if next_char.is_ascii_digit() {
+                        state = 2;
+                    } else {
+                        self.index = forward_lexeme - 1;
+                        return CoreToken::INTEGER;
+                    }
+                }
+                2 => {
+                    if next_char.is_ascii_digit() {
+                        // do nothing
+                    } else {
+                        self.index = forward_lexeme;
+                        return CoreToken::FLOATING_POINT_NUMBER;
+                    }
+                }
+                _ => {
+                    unreachable!("any state other than 0, 1, 2 and 3 is not reachable")
+                }
+            }
+            forward_lexeme += 1;
+        }
+        match state {
+            0 => {
+                self.index = forward_lexeme;
+                CoreToken::INTEGER
+            }
+            1 => {
+                self.index = forward_lexeme - 1;
+                CoreToken::INTEGER
+            }
+            2 => {
+                self.index = forward_lexeme;
+                CoreToken::FLOATING_POINT_NUMBER
+            }
+            _ => unreachable!("any state other than 0, 1, and 2 is not reachable"),
+        }
+    }
+
+    // : -> :, ::
+    pub fn extract_colon_prefix_lexeme(&mut self, code: &JarvilCode) -> CoreToken {
+        let forward_lexeme = self.index + 1;
+        if forward_lexeme < code.len() {
+            let next_char = code.get_char(forward_lexeme);
+            match next_char {
+                ':' => {
+                    self.index = forward_lexeme + 1;
+                    CoreToken::DOUBLE_COLON
+                }
+                _ => {
+                    self.index += 1;
+                    CoreToken::COLON
+                }
+            }
+        } else {
+            self.index += 1;
+            CoreToken::COLON
+        }
     }
 
     pub fn log_invalid_char_error(&mut self, token: &Token) {
