+++ compiler/src/lexer/mod.rs
@@ -1,5 +1,4 @@
 #[macro_use]
 pub mod macros;
-pub mod helper;
 pub mod lexer;
 pub mod token;
